WE-113 dragDrop 컴포넌트 onChange 함수 추상화 제거
컴포넌트 재사용성 높히기 위해 처리합니다!
