+++ pages/register/photo.tsx
@@ -1,14 +1,18 @@
 import { useRouter } from 'next/router';
-import { ReactElement, useEffect, useState } from 'react';
+import { ReactElement, useCallback } from 'react';
 import styled from 'styled-components';
 import Button from '../../components/common/Button';
 import DragDrop from '../../components/common/DragDrop';
 import Title from '../../components/common/Title';
 import LoginLayout from '../../components/layout/LoginLayout';
+import { useRegister } from '../../hooks/useRegister';
+import { encodeBase64ImageFile } from '../../library/ImageConverter';
 
 export default function ProfileImage() {
   const router = useRouter();
-  const [image, setImage] = useState('');
+  const [registerUser, setRegisterUser] = useRegister();
+
+  const { image_url: image } = registerUser;
 
   const handleNextButtonClick = () => {
     if (!image) {
@@ -17,22 +21,26 @@ export default function ProfileImage() {
     router.push('/register/introduction');
   };
 
-  useEffect(() => {
-    const persistNickname = localStorage.getItem('sign_up-nickname');
-    const persistImage = localStorage.getItem('sign_up-photo');
-    if (!persistNickname) {
-      router.replace('/register/nickname');
-    } else {
-      if (persistImage) {
-        setImage(persistImage);
+  const onChangeImage = useCallback(
+    async (e: React.ChangeEvent<HTMLInputElement> | any) => {
+      let selectedImage;
+      if (e.type === 'change') {
+        selectedImage = e.target.files[0];
+      } else if (e.type === 'drop') {
+        selectedImage = e.dataTransfer.files[0];
       }
-    }
-  }, []);
+      if (selectedImage && selectedImage.size <= 2000000) {
+        const image_url = await encodeBase64ImageFile(selectedImage as File);
+        setRegisterUser({ ...registerUser, image_url });
+      }
+    },
+    [registerUser],
+  );
 
   return (
     <Container>
       <Title>Please register your profile picture</Title>
-      <DragDrop image={image} setImage={setImage} />
+      <DragDrop image={image} onChangeImage={onChangeImage} />
       <StyledButton
         type="button"
         size="medium"
